requirements.txt
torch
transformers
peft
datasets
evaluate
matplotlib
seaborn
pandas
pyyaml
scikit-learn
nltk
rouge-score
sentencepiece